fix(blob-storage-database): move DatabaseBlobContent to Domain/
Architecture test 'Domain_types_should_reside_in_Domain_folder' requires
types inheriting from Entity / AuditedEntity / AggregateRoot / ValueObject
to live in a Domain/ subfolder. DatabaseBlobContent (CreationAuditedEntity)
was in Entities/ — relocate the entity and its tests, update the namespace
to Granit.BlobStorage.Database.Domain.
